requirements.txt
django
djangorestframework
stripe>=15.0.0
python-decouple